requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-tenants
psycopg2-binary